+++ lib/src/features/editor/presentation/editor_screen.dart
@@ -245,6 +245,7 @@ class _EditorScreenState extends ConsumerState<EditorScreen> {
             ),
           ],
         ),
+        floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
         floatingActionButton: isLoading
             ? Center(
                 child: PercentCircularIndicator(